fix: factorize TokenManagerHandlerMixin common class
